Renamed DataPointAdmission to Admission, Changed StratificationAnalysis to DataFilter, to better reflect it's usage.
Refactored members of Admission to clean up

Added loader for Box_Accession formatted box files, as well as a writer for the format. TODO - assign sample ids to older box formats, convert to Box_Accession.

SwabBag Accession and Box Accession are basically the same format, so the loader should essentialy be the same function.