Altered the way the Bin.Data property is handled.
Old way - would build a list of data points from the DataByPatientAdmissions table.

New way -
Backing variable data (List<DataPoint>)
Adds a "DataPropertyInvalid" flag, which is set to true only when new data is added, or data in the DataByPatientAdmissionsTable is set or changed (rare).
Rebuilds this list if this flag is set, otherwise returns the original list.
In the future, I may change DataPoint back to a Class, instead of a struct, to reduce the need for this kind of workaround.

Additionally - added a new filter - RemoveDataWithoutCDiffResult(Bin). Due to a large number of data points having invalid test results (recorded as NA in the c diff result column), it seems necessary to remove them. The bottlenecking this function created by calling Bin.Data was the reason for all of these changes.